backend/api/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
django-environ==0.8.1
django-freeze==0.6.4
idna==3.3
psycopg2-binary==2.9.3
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
xmltodict==0.12.0
django-simple-history
django-rest_framework
Pillow